[共通] fix warning

+++ Siv3D/src/Siv3D/ImageFormat/TGA/TGAEncoder.cpp
@@ -67,13 +67,13 @@ namespace s3d
 
 		const Color* pSrc = image[0];
 
-		for (size_t y = 0; y < height; ++y)
+		for (int32 y = 0; y < height; ++y)
 		{
 			std::memcpy(line.get(), pSrc, stride_bytes);
 
 			size_t index = 0;
 
-			for (size_t x = 0; x < width; ++x)
+			for (int32 x = 0; x < width; ++x)
 			{
 				std::swap(line[index], line[index + 2]);
 				index += 4;
@@ -91,7 +91,7 @@ namespace s3d
 	{
 		const int32 width			= image.width();
 		const int32 height			= image.height();
-		const size_t stride_bytes	= image.stride();
+		const uint32 stride_bytes	= image.stride();
 		const uint32 size_bytes		= (stride_bytes * height);
 		const TGAHeader header		= TGAHeader::Make(width, height);
 
@@ -112,13 +112,13 @@ namespace s3d
 		const Color* pSrc = image[0];
 		Byte* pDstLine = (blob.data() + sizeof(header));
 
-		for (size_t y = 0; y < height; ++y)
+		for (int32 y = 0; y < height; ++y)
 		{
 			std::memcpy(line.get(), pSrc, stride_bytes);
 
 			size_t index = 0;
 
-			for (size_t x = 0; x < width; ++x)
+			for (int32 x = 0; x < width; ++x)
 			{
 				std::swap(line[index], line[index + 2]);
 				index += 4;
